+++ packages/sdk/src/oracle/index.ts
@@ -1,1 +1,4 @@
+export * as client from "./client";
 export * as services from "./services";
+export * as types from "./types";
+export * as store from "./store";