+++ packages/contracts-node/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.18](https://github.com/UMAprotocol/protocol/compare/@uma/contracts-node@0.1.17...@uma/contracts-node@0.1.18) (2022-01-11)
+
+**Note:** Version bump only for package @uma/contracts-node
+
 ## [0.1.17](https://github.com/UMAprotocol/protocol/compare/@uma/contracts-node@0.1.16...@uma/contracts-node@0.1.17) (2021-12-17)
 
 **Note:** Version bump only for package @uma/contracts-node