+++ packages/optimistic-oracle/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.9.0](https://github.com/UMAprotocol/protocol/compare/@uma/optimistic-oracle@2.8.8...@uma/optimistic-oracle@2.9.0) (2022-01-11)
+
+### Features
+
+- **across-docs:** small nit relayer config ([#3759](https://github.com/UMAprotocol/protocol/issues/3759)) ([8384185](https://github.com/UMAprotocol/protocol/commit/8384185fdde23128d4f3a7b628fb6c6a11e30d51))
+
 ## [2.8.8](https://github.com/UMAprotocol/protocol/compare/@uma/optimistic-oracle@2.8.7...@uma/optimistic-oracle@2.8.8) (2021-12-17)
 
 ### Bug Fixes