+++ packages/scripts/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.6.0](https://github.com/UMAprotocol/protocol/compare/@uma/scripts@1.5.5...@uma/scripts@1.6.0) (2022-01-11)
+
+### Features
+
+- add manual relay script ([#3731](https://github.com/UMAprotocol/protocol/issues/3731)) ([ade5ad6](https://github.com/UMAprotocol/protocol/commit/ade5ad63718e15e83d4789ffe7b7814e31c6f07d))
+
 ## [1.5.5](https://github.com/UMAprotocol/protocol/compare/@uma/scripts@1.5.4...@uma/scripts@1.5.5) (2021-12-17)
 
 **Note:** Version bump only for package @uma/scripts