+++ packages/trader/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.9.0](https://github.com/UMAprotocol/protocol/compare/@uma/trader@1.8.8...@uma/trader@1.9.0) (2022-01-11)
+
+### Features
+
+- add transaction bundler to relayer for better batching ([#3723](https://github.com/UMAprotocol/protocol/issues/3723)) ([51902a8](https://github.com/UMAprotocol/protocol/commit/51902a8cfbbb60dc30b868c5fd3e9fd0f31d48b4))
+
 ## [1.8.8](https://github.com/UMAprotocol/protocol/compare/@uma/trader@1.8.7...@uma/trader@1.8.8) (2021-12-17)
 
 ### Bug Fixes